+++ changelog.md
@@ -4,6 +4,8 @@
   * `ClassFactory` added `isFinalObject() { return true/false }` to prevent addtional processing from happening if the `ClassFactory` creates the object AND assigns all values.
   * Fixed an issue with classes that used custom reader/writes being loaded when inside an array or collection. If there was circular references, they were not resolved correctly.
   * This version writes `Enums` in a more compact way with the field name associated to a JSON String name of the enum.  However, the prior versions of `json-io` wrote `Enums` out as JSON objects.  The JSON reader will read `Enums` either way.  If you want the output to continue to write `Enums` as a JSON Object, use the `.writeEnumsAsOptions()` on the `WriteOptionsBuilder`, and it will output enums as it used to.
+  * Minor change: `JsonObject` was `JsonObject<K, V>` and is now `JsonObject` (no generics).  If you used `JsonObject` in your code, make sure to remove the generics.
+  * Minor change: `JsonReader.ClassFactory::newInstance(Class c, Object)` has been changed to `JsonReader.ClassFactory::newInstance(Class<?>, JsonObject)`.  If you have written a `CustomClassFactory,` update the method signature to `newInstance(Class<?>, JsonObject).` 
 * 4.14.2
   * Enum/EnumSet support fully added @kpartlow
   * `WARN` This version inadvertently slipped to `JDK11+` (which has been corrected in `4.15.0`).  Version `5.x.x` will be `JDK11 or JDK17`.